+++ src/Webhooks/ZadarmaWebhookEvent.php
@@ -6,17 +6,17 @@ namespace GracjanKubicki\LaravelZadarma\Webhooks;
 
 enum ZadarmaWebhookEvent: string
 {
-    case NotifyStart = 'notify_start';
-    case NotifyInternal = 'notify_internal';
-    case NotifyAnswer = 'notify_answer';
-    case NotifyEnd = 'notify_end';
-    case NotifyOutStart = 'notify_out_start';
-    case NotifyOutEnd = 'notify_out_end';
-    case NotifyRecord = 'notify_record';
-    case NotifyIvr = 'notify_ivr';
-    case NumberLookup = 'number_lookup';
-    case CallTracking = 'call_tracking';
-    case Sms = 'sms';
-    case SpeechRecognition = 'speech_recognition';
-    case Document = 'document';
+    case NotifyStart = 'NOTIFY_START';
+    case NotifyInternal = 'NOTIFY_INTERNAL';
+    case NotifyAnswer = 'NOTIFY_ANSWER';
+    case NotifyEnd = 'NOTIFY_END';
+    case NotifyOutStart = 'NOTIFY_OUT_START';
+    case NotifyOutEnd = 'NOTIFY_OUT_END';
+    case NotifyRecord = 'NOTIFY_RECORD';
+    case NotifyIvr = 'NOTIFY_IVR';
+    case NumberLookup = 'NUMBER_LOOKUP';
+    case CallTracking = 'CALL_TRACKING';
+    case Sms = 'SMS';
+    case SpeechRecognition = 'SPEECH_RECOGNITION';
+    case Document = 'DOCUMENT';
 }